backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=24.1.0
Pillow>=11.0.0
opencv-python-headless>=4.10.0.84
torch
torchvision
ultralytics>=8.3.0
numpy>=2.1.0
scipy>=1.14.0
matplotlib>=3.9.0
reportlab>=4.2.0
pandas>=2.2.0
websockets>=13.0
httpx>=0.27.0
python-dotenv>=1.0.1
psutil>=6.0.0